Revert "updated_utils"
